server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0+cpu
torchvision==0.16.0+cpu
--extra-index-url https://pypi.org/simple
numpy==1.24.3
pillow==10.2.0
google-generativeai==0.3.2
python-dotenv==1.0.1 